crowd_behaviour_forecasting/requirements.txt
# Core Dependencies
torch==2.1.0
torchvision==0.16.0
torch-geometric==2.3.0
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
scipy==1.11.2

# Video Processing & MOT
ultralytics==8.0.185
motmetrics==1.2.0
lap==0.4.0
cython-bbox==0.1.3

# Transformers & Deep Learning
transformers==4.33.0
timm==0.9.7
einops==0.7.0

# Edge Deployment & Model Optimization
onnx==1.14.1
onnxruntime==1.16.0
onnxruntime-gpu==1.16.0
tensorrt==8.6.0
quantization-aware-training==0.1.0

# Web & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
websockets==11.0.3

# Visualization & Analysis
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
tensorboard==2.14.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
Pillow==10.0.1
requests==2.31.0

# Testing & Quality
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
mypy==1.6.1

# Logging & Monitoring
loguru==0.7.2

# Docker (optional but recommended)
docker==6.1.0
